Move components2 to src root
