+++ lib/src/main/java/com/zetcode/components2/TextAreaEx.java
@@ -1,0 +1,63 @@
+package com.zetcode.components2;
+
+import javax.swing.GroupLayout;
+import javax.swing.JComponent;
+import javax.swing.JFrame;
+import javax.swing.JScrollPane;
+import javax.swing.JTextArea;
+import java.awt.Dimension;
+import java.awt.EventQueue;
+
+public class TextAreaEx extends JFrame {
+
+    public TextAreaEx() {
+
+        initUI();
+    }
+
+    private void initUI() {
+
+        var area = new JTextArea();
+        var spane = new JScrollPane(area);
+
+        area.setLineWrap(true);
+        area.setWrapStyleWord(true);
+
+        createLayout(spane);
+
+        setTitle("JTextArea");
+        setSize(new Dimension(350, 300));
+        setDefaultCloseOperation(EXIT_ON_CLOSE);
+        setLocationRelativeTo(null);
+    }
+
+    private void createLayout(JComponent... arg) {
+
+        var pane = getContentPane();
+        var gl = new GroupLayout(pane);
+        pane.setLayout(gl);
+
+        gl.setAutoCreateContainerGaps(true);
+        gl.setAutoCreateGaps(true);
+
+        gl.setHorizontalGroup(gl.createParallelGroup()
+                .addComponent(arg[0])
+
+        );
+
+        gl.setVerticalGroup(gl.createSequentialGroup()
+                .addComponent(arg[0])
+        );
+
+        pack();
+    }
+
+    public static void main(String[] args) {
+
+        EventQueue.invokeLater(() -> {
+
+            var ex = new TextAreaEx();
+            ex.setVisible(true);
+        });
+    }
+}
