+++ lib/src/main/java/com/zetcode/components2/med.png
