+++ lib/src/main/java/com/zetcode/components2/min.png
